+++ events/client/message.js
@@ -1,0 +1,18 @@
+const {PREFIX} = require('../../config.js');
+
+module.exports = (client, msg) => {
+    // A Corriger : uniquement si début du message
+    // if (msg.mentions.has(client.user.id)) {
+    //     return msg.reply(`Tu as besoin d'aide ? Mon préfixe est \`${PREFIX}\``);
+    // }
+
+    if(!msg.content.startsWith(PREFIX) || msg.author.bot || msg.channel.type === "dm") return;
+
+    const args = msg.content.slice(PREFIX.length).split(/ +/);
+    const commandName = args.shift().toLowerCase();
+
+    const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.help.aliases && cmd.help.aliases.includes(commandName));
+    if(!command) return;
+
+    command.run(client, msg, args);
+}