+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 ## [v0.1.0]
 ### Added
 - Commands :
-    - help
+    - help (aliases: h)
+    - info (aliases: bot, botinfo, about)
+    - user (aliases: userinfo)
+    - serverinfo (aliases: si)
+    - uptime (aliases: up)
+    - ping
 ## Creation of the project on January 17, 2021